+++ data-migrations/2026-04-29-1432-create-alternateIdentifiers.sql
@@ -1,17 +1,13 @@
--- Join table to store relationship between plans and guidance affiliations per userId
+-- Join table to store relationship between plans and alternate identifiers
 CREATE TABLE `alternateIdentifiers` (
   `id` INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
   `planId` INT UNSIGNED NOT NULL,
   `alternateIdentifier` VARCHAR(512) NOT NULL,
-  `userId` INT UNSIGNED NOT NULL,
   `created` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
   `createdById` int NOT NULL,
   `modified` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
   `modifiedById` int NOT NULL,
 
-  UNIQUE KEY `unique_plan_affiliation_user` (`planId`, `affiliationId`, `userId`),
-
   FOREIGN KEY (`planId`) REFERENCES `plans`(`id`) ON DELETE CASCADE,
-  FOREIGN KEY (`affiliationId`) REFERENCES `affiliations`(`uri`) ON DELETE CASCADE,
-  FOREIGN KEY (`userId`) REFERENCES `users`(`id`) ON DELETE CASCADE
+  UNIQUE KEY `unique_planAlternateIds` (`planId`,`alternateIdentifier`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;