+++ src/schemas/plan.ts
@@ -7,6 +7,9 @@ export const typeDefs = gql`
 
     "Get a specific plan"
     plan(planId: Int!): Plan
+
+    "Lookup a plan by an alternate identifier"
+    planByAlternateIdentifier(alternateIdentifier: String!): Plan
   }
 
   extend type Mutation {